refactor(danmaku-view-hook): move danmaku keep-visible feature under ui scope
