+++ app/src/main/java/io/github/twyora/douyinenhancer/config/key/FeedKey.kt
@@ -4,5 +4,4 @@ object FeedKey {
     const val BYPASS_LISTEN_AWEME_RESTRICTION = "bypass_listen_aweme_restriction"
     const val FEED_DOUBLE_TAP_DIGG = "disable_feed_double_tap_digg"
     const val FEED_DOUBLE_TAP_OPEN_COMMENT = "feed_double_tap_open_comment"
-    const val KEEP_DANMAKU_VISIBLE = "keep_danmaku_visible"
 }
