+++ app/src/main/java/io/github/twyora/douyinenhancer/config/key/UiKey.kt
@@ -1,0 +1,5 @@
+package io.github.twyora.douyinenhancer.config.key
+
+object UiKey {
+    const val KEEP_DANMAKU_VISIBLE = "keep_danmaku_visible"
+}
