+++ app/src/main/res/values/strings.xml
@@ -27,8 +27,8 @@
     <string name="recommended_feed_filter_dialog_title">Recommended Feed Filter</string>
     <string name="pref_feed_bypass_listen_aweme_restriction_title">Bypass Listen Aweme Restriction</string>
     <string name="pref_feed_bypass_listen_aweme_restriction_summary">Skip the audio check when entering listen aweme mode</string>
-    <string name="pref_feed_keep_danmaku_visible_title">Keep Danmaku Visible</string>
-    <string name="pref_feed_keep_danmaku_visible_summary">Keep danmaku visible when entering clean mode</string>
+    <string name="pref_ui_keep_danmaku_visible_title">Keep Danmaku Visible</string>
+    <string name="pref_ui_keep_danmaku_visible_summary">Keep danmaku visible when entering clean mode</string>
     <string name="recommended_feed_filter_main_switch_title">Main switch</string>
     <string name="recommended_feed_filter_main_switch_summary">Turn on to use recommended feed filtering features</string>
     <string name="recommended_feed_filter_block_ad_title">✦ Block ads</string>
@@ -76,6 +76,7 @@
     <string name="config_import_chooser">Select configuration</string>
 
     <string name="pref_category_misc">Misc</string>
+    <string name="pref_category_ui">UI</string>
     <string name="pref_misc_enable_hidden_features_title">✦ Enable hidden features</string>
     <string name="pref_misc_enable_hidden_features_summary">Enable to use hidden features</string>
     <string name="pref_misc_enable_hidden_features_steps_remaining">Perform %1$d more steps to enable hidden features</string>
